+++ pyscm/tests/utility_tests.py
@@ -8,7 +8,6 @@ from unittest import TestCase
 from sklearn.utils import estimator_checks
 
 from .._scm_utility import find_max
-from ..scm import SetCoveringMachineClassifier
 
 def eprint(*args, **kwargs):
     print(*args, file=sys.stderr, **kwargs)
@@ -113,7 +112,7 @@ class UtilityTests(TestCase):
         best_thresholds, best_kinds = find_max(p, X, y, Xas, np.array([1, 2], dtype=np.int), np.ones(X.shape[1]))
         np.testing.assert_almost_equal(actual=best_feat_idx, desired=[0, 1])
 
-    def automatic_testing(self):
+    def test_random_data(self):
         """
         Random testing
         """
@@ -155,44 +154,3 @@ class UtilityTests(TestCase):
 
                     np.testing.assert_almost_equal(actual=solver_best_utility, desired=max(max(le_rule_utilities),
                                                                                            max(g_rule_utilities)))
-
-    def sklearn_compatibility_test(self):
-        """
-        Test Sklearn compatibility
-
-        """
-        for check in _yield_check_sklearn_compatibility():
-            check
-
-
-def _yield_check_sklearn_compatibility():
-    """
-    Assert Sklearn compatibility. It is fully compatible except it does not handle multi-class.
-    Uses sklearn test suits.
-    If fails, will raise an exception
-    """
-    yield estimator_checks.check_parameters_default_constructible("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
-    for check in estimator_checks._yield_non_meta_checks("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier):
-        yield check
-
-    for check in estimator_checks._yield_classifier_checks("SetCoveringMachineClassifier",
-                                                           SetCoveringMachineClassifier):
-        yield check
-
-    # Will fail because SCM does not handle multi-class.
-    #yield estimator_checks.check_fit2d_predict1d("SetCoveringMachineClassifier",
-    #                                                     SetCoveringMachineClassifier)
-    yield estimator_checks.check_fit2d_1sample("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
-    yield estimator_checks.check_fit2d_1feature("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
-    yield estimator_checks.check_fit1d_1feature("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
-    yield estimator_checks.check_fit1d_1sample("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
-    yield estimator_checks.check_get_params_invariance("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
-    yield estimator_checks.check_dict_unchanged("SetCoveringMachineClassifier",
-                                                         SetCoveringMachineClassifier)
